Updates for sharedArtifactDirectory reporting issue

+++ src/main/resources/scripts/copy_build_dir.py
@@ -28,18 +28,41 @@ import subprocess
 import tarfile
 import sqlite3
 import shutil
+import argparse
 
+try:
+    ## This tests to see if 2018 is present.
+    from vector.apps.ReportBuilder.custom_report import CustomReport
+    try:
+        from vector.apps.DataAPI.unit_test_api import UnitTestApi
+    except:
+        from vector.apps.DataAPI.api import Api as UnitTestApi
+except:
+    pass
+    
+try:
+    from vector.apps.DataAPI.vcproject_api import VCProjectApi
+except:
+    pass
 
-def make_relative(path, workspace):
+def make_relative(path, workspace, vCastProjectWorkspace, ManageProjectName):
 
     path = path.replace("\\","/")
     
+
+    if path.endswith('.LIS') or '000000' in path:
+        return path
+
     if not os.path.isabs(path):
         return path
         
     # if the paths match
     if path.lower().startswith(workspace.lower()):
-        path = path[len(workspace)+1:]
+        if "@" in path:
+            path = path[len(workspace):]
+            path = path.split("/",1)[1]
+        else:
+            path = path[len(workspace)+1:]
         
     # if the paths match except for first character (d:\ changed to j:\)
     elif path.lower()[1:].startswith(workspace.lower()[1:]):
@@ -51,17 +74,21 @@ def make_relative(path, workspace):
         
         path = path[workspaceIndex:].split("/",2)[2]
         
+    elif vCastProjectWorkspace is not None and vCastProjectWorkspace.lower() in path.lower():
+        path = path.replace(vCastProjectWorkspace,"")
+        path = ManageProjectName[:-4].replace("\\", "/") + path
+        
     else:
         print("  Warning: Unable to convert source file: " + path + " to relative path based on WORKSPACE: " + workspace)
-
+    
     return path
 
     
-def updateDatabase(conn, nocase, workspace, updateWhat, updateFrom):
+def updateDatabase(conn, nocase, workspace, updateWhat, updateFrom, vCastProjectWorkspace, ManageProjectName):
     sql = "SELECT id, %s FROM %s" % (updateWhat, updateFrom)
     files = conn.execute(sql)
     for id_, path in files:
-        relative = make_relative(path,workspace)    
+        relative = make_relative(path,workspace, vCastProjectWorkspace, ManageProjectName)    
         sql = "UPDATE %s SET %s = '%s' WHERE id=%s COLLATE NOCASE" % (updateFrom, updateWhat, relative, id_)
         conn.execute(sql)
         
@@ -93,11 +120,12 @@ def addDirectory(tf, build_dir, dir):
             for fname in fileList:
                 tf.add(os.path.join(dirName, fname))
 
-def addConvertCoverFile(tf, file, workspace, build_dir, nocase):
-
-    print("Updating cover.db")
+def addConvertCoverFile(tf, file, workspace, build_dir, nocase, vCastProjectWorkspace, ManageProjectName, noTar):
+    
     fullpath = build_dir + os.path.sep + file
     bakpath = fullpath + '.bk'
+
+    print("Updating cover.db: ", fullpath)
     
     if os.path.isfile(fullpath):
         conn = sqlite3.connect(fullpath)
@@ -106,19 +134,20 @@ def addConvertCoverFile(tf, file, workspace, build_dir, nocase):
 
             # update the database paths to be relative from workspace
             try:
-                updateDatabase(conn, nocase, workspace, "LIS_file", "instrumented_files")
+                updateDatabase(conn, nocase, workspace, "LIS_file", "instrumented_files", vCastProjectWorkspace, ManageProjectName)
             except:
-                updateDatabase(conn, nocase, workspace, "path", "lis_files")
-            updateDatabase(conn, nocase, workspace, "display_path", "source_files")
-            updateDatabase(conn, nocase, workspace, "path", "source_files")
+                updateDatabase(conn, nocase, workspace, "path", "lis_files" , vCastProjectWorkspace, ManageProjectName)
+            updateDatabase(conn, nocase, workspace, "display_path", "source_files", vCastProjectWorkspace, ManageProjectName)
+            updateDatabase(conn, nocase, workspace, "path", "source_files", vCastProjectWorkspace, ManageProjectName)
             
             conn.commit()
             conn.close()
             addFile(tf, file, build_dir)
-            os.remove(fullpath)
-            shutil.move(bakpath, fullpath)
+            if not noTar:
+                os.remove(fullpath)
+                shutil.move(bakpath, fullpath)
 
-def addConvertMasterFile(tf, file, workspace, build_dir, nocase):
+def addConvertMasterFile(tf, file, workspace, build_dir, nocase, vCastProjectWorkspace, ManageProjectName, noTar):
     print("Updating master.db")
     fullpath = build_dir + os.path.sep + file
     bakpath = fullpath + '.bk'
@@ -126,21 +155,24 @@ def addConvertMasterFile(tf, file, workspace, build_dir, nocase):
         conn = sqlite3.connect(fullpath)
         if conn:
             shutil.copyfile(fullpath, bakpath)
-            updateDatabase(conn, nocase, workspace, "path", "sourcefiles")
+            updateDatabase(conn, nocase, workspace, "path", "sourcefiles", vCastProjectWorkspace, ManageProjectName)
             conn.commit()
             conn.close()
             addFile(tf, file, build_dir)
-            os.remove(fullpath)
-            shutil.move(bakpath, fullpath)
+            if not noTar:
+                os.remove(fullpath)
+                shutil.move(bakpath, fullpath)
 
-def addConvertFiles(tf, workspace, build_dir, nocase):
-    addConvertCoverFile (tf, "cover.db", workspace, build_dir, nocase)
-    addConvertMasterFile(tf, "master.db", workspace, build_dir, nocase)
+def addConvertFiles(tf, workspace, build_dir, nocase, vCastProjectWorkspace, ManageProjectName, noTar):
+    addConvertCoverFile (tf, "cover.db",  workspace, build_dir, nocase, vCastProjectWorkspace, ManageProjectName, noTar)
+    addConvertMasterFile(tf, "master.db", workspace, build_dir, nocase, vCastProjectWorkspace, ManageProjectName, noTar)
 
-def run(ManageProjectName, Level, BaseName, Env, workspace):
+
+def run(ManageProjectName, Level, BaseName, Env, workspace, vCastProjectWorkspace, noTar):
 
     if sys.platform.startswith('win32'):
         workspace = workspace.replace("\\", "/")
+        vCastProjectWorkspace = vCastProjectWorkspace.replace("\\", "/")
         nocase = "COLLATE NOCASE"
     else:
         nocase = ""
@@ -168,9 +200,12 @@ def run(ManageProjectName, Level, BaseName, Env, workspace):
 
     if build_dir != "":
         build_dir = build_dir + os.path.sep + Env
-        tf = tarfile.open(BaseName + "_build.tar", mode='w')
+        if noTar:
+            tf = tarfile.open("delete_me_" + BaseName + "_build.tar", mode='w')
+        else:
+            tf = tarfile.open(BaseName + "_build.tar", mode='w')
         try:
-            addConvertFiles(tf, workspace, build_dir, nocase)
+            addConvertFiles(tf, workspace, build_dir, nocase, vCastProjectWorkspace, ManageProjectName, noTar)
             addFile(tf, "testcase.db", build_dir)
             addFile(tf, "COMMONDB.VCD", build_dir)
             addFile(tf, "UNITDATA.VCD", build_dir)
@@ -195,14 +230,39 @@ def run(ManageProjectName, Level, BaseName, Env, workspace):
 
         finally:
             tf.close()
+            
+        if noTar:
+            os.remove("delete_me_" + BaseName + "_build.tar")
 
+
+def getVcastProjectWorkspace(args):
+
+    with VCProjectApi(args.ManageProject) as vc_api:
+#    vc_api = VCProjectApi(args.ManageProject)
+        vCastProjectWorkspace = vc_api.project.workspace
     
+    return vCastProjectWorkspace
+
 if __name__ == '__main__':
-                
-    ManageProjectName = sys.argv[1]
-    Level = sys.argv[2]
-    BaseName = sys.argv[3]
-    Env = sys.argv[4]
+    
+    
+    parser = argparse.ArgumentParser()
+    parser.add_argument('ManageProject',       help='Manager Project Name')
+    parser.add_argument('-l', '--level',       help='Environment Name if only doing single environment.  Should be in the form of level/env', default="NotProvided/NotProvided")
+    parser.add_argument('-b', '--basename',    help='Enable verbose output',    default="")
+    parser.add_argument('-e', '--environment', help='Enable verbose output', default="")
+    parser.add_argument('--notar',             help='Don\'t Product a tar file', default=False, action="store_true")
+    parser.add_argument('-v', '--verbose',     help='Enable verbose output',     action="store_true")
+    
+    args = parser.parse_args()
+    
+    ManageProjectName = args.ManageProject
+    Level = args.level
+    BaseName = args.basename
+    Env = args.environment
+    vCastProjectWorkspace = getVcastProjectWorkspace(args)
+    noTar = args.notar
+    
     workspace = os.getenv("WORKSPACE")
 
     if workspace is None:
@@ -213,5 +273,5 @@ if __name__ == '__main__':
 
     os.environ['VCAST_MANAGE_PROJECT_DIRECTORY'] = os.path.abspath(ManageProjectName).rsplit(".",1)[0]
     
-    run(ManageProjectName, Level, BaseName, Env, workspace)
+    run(ManageProjectName, Level, BaseName, Env, workspace, vCastProjectWorkspace, noTar)
